Fix bugs in the arithmetic compressor and CLI
The entire compression interface was horribly broken, from
typos, to incorrect itertools chaining, to docopt inconsitencies,
to a broken implementation of arithmetic compression.

Quick comment on the last issue.  Previously it tried to be smart
and use the probability returned by `model.update` to compute the
predicted probability of a zero.  But when the next symbol was a 1
this required a subtraction to compute the probability of a zero
which was just enough in edge cases to get a different rounding of
the floating point errors than was used in decoding  Ack!!

To recreate this issue compress [1, 1, 1, 0] with a KT model using
4-bits in the arithmetic compressor.  Compressing the third bit
created exactly an integral split point which would round differently
in the compressor and decompressor.